MOD: prueba

+++ vtex/orders_api/ORDERS/descomposicion_orders/item.py
@@ -335,7 +335,7 @@ def run():
         "mode": "NULLABLE"
     },{
         "name": "commission",
-        "type": "INTEGER",
+        "type": "FLOAT",
         "mode": "NULLABLE"
     },{
         "name": "productId",
